chore: address Bugbot review on onboarding UI polish
Remove unused NewAppModal and filterAppCatalog after migrating to the
/apps/new page, and extract CanvasCardDescription from CanvasCardsGrid.
